[ticket/62558] Support reversal of multi inline-edits on same find
With this change, the reverse is now built just like the normal
(install) action - per the process_edits method of the acp_mods
class. This fixes the bug where AM failed to reverse all but
the last one of the multiple in-line edit's on the same line/find.
Also did some minor optimizations of editor->build_uninstall.