[ticket/47075] Clean-up installation/uninstallation of contrib MODx
An additional MODx file that has been force installed will cause us
to have an empty $action_ary when uninstalling it. Additionally we
need to make sure to send the parent ID and the link to the MODx file
via hidden fields in order to be able to force uninstall those MODx
files.